+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 - Fix layout rendering when loading a flow (#29)
 - Use label instead of stepId for naming (#30)
 - Support keyboard shortcuts (save, delete, copy, paste) (#31)
-- Organize sidebar with accordions (#32)
+- Organize sidebar with accordions and spinner (#32)
 
 ## 0.0.6